#153: 신한 api dto 네이밍 변경 및 위치 변경
