+++ src/main/java/com/ssafy/ssashinsa/heyfy/shinhanApi/dto/account/exchange/ShinhanAccountInfoRecDto.java
@@ -1,11 +1,11 @@
-package com.ssafy.ssashinsa.heyfy.exchange.dto.external.shinhan;
+package com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.exchange;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
-public class AccountInfoRecDto {
+public class ShinhanAccountInfoRecDto {
     private String accountNo;
     private Double amount;
     private Double balance;
