+++ src/main/java/com/ssafy/ssashinsa/heyfy/shinhanApi/dto/account/exchange/ShinhanExchangeCurrencyRecDto.java
@@ -1,4 +1,4 @@
-package com.ssafy.ssashinsa.heyfy.exchange.dto.external.shinhan;
+package com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.exchange;
 
 import lombok.Builder;
 import lombok.Data;
