+++ src/main/java/com/ssafy/ssashinsa/heyfy/shinhanApi/dto/account/exchange/ShinhanExchangeResponseDto.java
@@ -1,18 +1,15 @@
-package com.ssafy.ssashinsa.heyfy.exchange.dto.external.shinhan;
+package com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.exchange;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.common.ShinhanCommonRequestHeaderDto;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AccountBalanceRequestDto {
+public class ShinhanExchangeResponseDto {
     @JsonProperty("Header")
     private ShinhanCommonRequestHeaderDto Header;
-    private String accountNo;
+    @JsonProperty("REC")
+    private ShinhanExchangeResponseRecDto REC;
 }
