+++ src/main/java/com/ssafy/ssashinsa/heyfy/shinhanApi/dto/account/inquire/ShinhanInquireDemandDepositAccountBalanceRequestDto.java
@@ -1,15 +1,18 @@
-package com.ssafy.ssashinsa.heyfy.exchange.dto.external.shinhan;
+package com.ssafy.ssashinsa.heyfy.shinhanApi.dto.account.inquire;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.ssafy.ssashinsa.heyfy.shinhanApi.dto.common.ShinhanCommonRequestHeaderDto;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
-public class ShinhanExchangeResponseDto {
+@NoArgsConstructor
+@AllArgsConstructor
+public class ShinhanInquireDemandDepositAccountBalanceRequestDto {
     @JsonProperty("Header")
     private ShinhanCommonRequestHeaderDto Header;
-    @JsonProperty("REC")
-    private ShinhanExchangeResponseRecDto REC;
+    private String accountNo;
 }
